// src/firestore/types/firestore-query-operations.types.ts
export enum FirestoreQueryOperation {
    LessThan = "<",
    LessThanOrEqualTo = "<=",
    EqualTo = "==",
    GreaterThanOrEqualTo = ">=",
    GreaterThan = ">",
    NotEqualTo = "!=",
    In = "in",
    NotIn = "not-in",
    ArrayContains = "array-contains",
    ArrayContainsAny = "array-contains-any"
}